requirements.txt
fastapi
uvicorn
pymongo
chromadb
python-dotenv
pydantic
sentence-transformers
google-generativeai
groq
reportlab